+++ lib/Makefile.uk
@@ -61,7 +61,7 @@ $(eval $(call import_lib,$(CONFIG_UK_BASE)/lib/uksglist))
 $(eval $(call import_lib,$(CONFIG_UK_BASE)/lib/uksignal))
 $(eval $(call import_lib,$(CONFIG_UK_BASE)/lib/uksp))
 $(eval $(call import_lib,$(CONFIG_UK_BASE)/lib/ukstore))
-$(eval $(call import_lib,$(CONFIG_UK_BASE)/lib/ukswrand))
+$(eval $(call import_lib,$(CONFIG_UK_BASE)/lib/ukrandom))
 $(eval $(call import_lib,$(CONFIG_UK_BASE)/lib/uktest))
 $(eval $(call import_lib,$(CONFIG_UK_BASE)/lib/posix-time))
 $(eval $(call import_lib,$(CONFIG_UK_BASE)/lib/uktimeconv))
