+++ lib/ukrandom/Makefile.uk
@@ -1,0 +1,12 @@
+$(eval $(call addlib_s,libukrandom,$(CONFIG_LIBUKRANDOM)))
+
+CINCLUDES-$(CONFIG_LIBUKRANDOM)	+= -I$(LIBUKRANDOM_BASE)/include
+CXXINCLUDES-$(CONFIG_LIBUKRANDOM) += -I$(LIBUKRANDOM_BASE)/include
+
+LIBUKRANDOM_SRCS-$(CONFIG_LIBUKRANDOM_MWC) += $(LIBUKRANDOM_BASE)/mwc.c
+LIBUKRANDOM_SRCS-$(CONFIG_LIBUKRANDOM_CHACHA) += $(LIBUKRANDOM_BASE)/chacha.c
+LIBUKRANDOM_SRCS-$(CONFIG_LIBUKRANDOM_DEVFS) += $(LIBUKRANDOM_BASE)/dev.c
+LIBUKRANDOM_SRCS-y += $(LIBUKRANDOM_BASE)/random.c
+LIBUKRANDOM_SRCS-y += $(LIBUKRANDOM_BASE)/getrandom.c
+
+UK_PROVIDED_SYSCALLS-$(CONFIG_LIBUKRANDOM) += getrandom-3
