+++ lib/ukrandom/include/uk/random.h
@@ -1,6 +1,8 @@
 /* SPDX-License-Identifier: BSD-3-Clause */
 /*
- * Copyright (c) 2019, University Politehnica of Bucharest. All rights reserved.
+ * Authors: Simon Kuenzer <simon.kuenzer@neclab.eu>
+ *
+ * Copyright (c) 2018, NEC Europe Ltd., NEC Corporation. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -27,70 +29,48 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-#include <string.h>
-#include <uk/swrand.h>
+
+#ifndef __UK_RANDOM_H__
+#define __UK_RANDOM_H__
+
+#include <sys/types.h>
+#include <uk/arch/types.h>
+#include <uk/plat/lcpu.h>
 #include <uk/config.h>
-#include <uk/print.h>
-#include <uk/init.h>
+#include <uk/plat/time.h>
 
-__u32 uk_swrandr_gen_seed32(void)
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+struct uk_swrand;
+
+extern struct uk_swrand uk_swrand_def;
+
+void uk_swrand_init_r(struct uk_swrand *r, unsigned int seedc,
+			const __u32 seedv[]);
+__u32 uk_swrand_randr_r(struct uk_swrand *r);
+
+__u32 uk_swrandr_gen_seed32(void);
+/* Uses the pre-initialized default generator  */
+/* TODO: Add assertion when we can test if we are in interrupt context */
+/* TODO: Revisit with multi-CPU support */
+static inline __u32 uk_swrand_randr(void)
 {
-	__u32 val;
+	unsigned long iflags;
+	__u32 ret;
 
-#ifdef CONFIG_LIBUKSWRAND_INITIALSEED_TIME
-	val = (__u32)ukplat_wall_clock();
-#endif
+	iflags = ukplat_lcpu_save_irqf();
+	ret = uk_swrand_randr_r(&uk_swrand_def);
+	ukplat_lcpu_restore_irqf(iflags);
 
-#ifdef CONFIG_LIBUKSWRAND_INITIALSEED_RDRAND
-	asm volatile ("rdrand %%eax;"
-		: "=a" (val));
-#endif
-
-#ifdef CONFIG_LIBUKSWRAND_INITIALSEED_USECONSTANT
-	val = CONFIG_LIBUKSWRAND_INITIALSEED_CONSTANT;
-#endif
-
-	return val;
+	return ret;
 }
 
-ssize_t uk_swrand_fill_buffer(void *buf, size_t buflen)
-{
-	size_t step, chunk_size, i;
-	__u32 rd;
+ssize_t uk_swrand_fill_buffer(void *buf, size_t buflen);
 
-	step = sizeof(__u32);
-	chunk_size = buflen % step;
-
-	for (i = 0; i < buflen - chunk_size; i += step)
-		*(__u32 *)((char *) buf + i) = uk_swrand_randr();
-
-	/* fill the remaining bytes of the buffer */
-	if (chunk_size > 0) {
-		rd = uk_swrand_randr();
-		memcpy(buf + i, &rd, chunk_size);
-	}
-
-	return buflen;
+#ifdef __cplusplus
 }
-
-static int _uk_swrand_init(struct uk_init_ctx *ictx __unused)
-{
-	unsigned int i;
-#ifdef CONFIG_LIBUKSWRAND_CHACHA
-	unsigned int seedc = 10;
-	__u32 seedv[10];
-#else
-	unsigned int seedc = 2;
-	__u32 seedv[2];
 #endif
-	uk_pr_info("Initialize random number generator...\n");
 
-	for (i = 0; i < seedc; i++)
-		seedv[i] = uk_swrandr_gen_seed32();
-
-	uk_swrand_init_r(&uk_swrand_def, seedc, seedv);
-
-	return seedc;
-}
-
-uk_early_initcall(_uk_swrand_init, 0x0);
+#endif /* __UK_RANDOM_H__ */
