--- lib/ukswrand/Makefile.uk
@@ -1,12 +1,0 @@
-$(eval $(call addlib_s,libukswrand,$(CONFIG_LIBUKSWRAND)))
-
-CINCLUDES-$(CONFIG_LIBUKSWRAND)	+= -I$(LIBUKSWRAND_BASE)/include
-CXXINCLUDES-$(CONFIG_LIBUKSWRAND) += -I$(LIBUKSWRAND_BASE)/include
-
-LIBUKSWRAND_SRCS-$(CONFIG_LIBUKSWRAND_MWC) += $(LIBUKSWRAND_BASE)/mwc.c
-LIBUKSWRAND_SRCS-$(CONFIG_LIBUKSWRAND_CHACHA) += $(LIBUKSWRAND_BASE)/chacha.c
-LIBUKSWRAND_SRCS-$(CONFIG_LIBUKSWRAND_DEVFS) += $(LIBUKSWRAND_BASE)/dev.c
-LIBUKSWRAND_SRCS-y += $(LIBUKSWRAND_BASE)/swrand.c
-LIBUKSWRAND_SRCS-y += $(LIBUKSWRAND_BASE)/getrandom.c
-
-UK_PROVIDED_SYSCALLS-$(CONFIG_LIBUKSWRAND) += getrandom-3
